requirements.txt
openai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
pytesseract>=0.3.10
opencv-python>=4.8.0 